feat: [speek] zustand 추가 및 파일정리
