+++ src/app/games/speaking/page.tsx
@@ -1,8 +1,10 @@
-import Speak from '../_components/Speak';
+import Speak from './_components/Speak';
+import Timer from './_components/Timer';
 
 const page = () => {
   return (
     <div>
+      <Timer></Timer>
       <Speak />
     </div>
   );
